requirements.txt
# Requirements.txt

# Pytube for YouTube video extraction
pytube==11.0.1

# Pandas for data manipulation
pandas==1.3.3

# Google API Client for YouTube Data API
google-api-python-client==2.15.0

# Google OAuth2 for handling credentials
google-auth-oauthlib==0.4.6